bugfix/setup-linter fix linter problems after back merge
