+++ ClientApp/src/components/SimpleDataTest.jsx
@@ -1,11 +1,12 @@
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
+// eslint-disable-next-line no-unused-vars
 function SimpleDataTest(props) {
   const [state, setState] = useState({ data: [], loading: true });
 
   const populateData = async () => {
     const data = await (await fetch("simpledatatest")).json();
-    setState({ data: data, loading: false });
+    setState({ data, loading: false });
   };
 
   useEffect(() => {
@@ -19,7 +20,7 @@ function SimpleDataTest(props) {
         This component gets data from a server API which itself gets data from a
         PostgreSQL database.
       </p>
-      <button className="btn btn-primary" onClick={populateData}>
+      <button type="button" className="btn btn-primary" onClick={populateData}>
         Reload
       </button>
       {state.loading ? (
